+++ sdk/ai/Azure.AI.Projects/src/Custom/Datasets/Datasets.cs
@@ -51,13 +51,19 @@ namespace Azure.AI.Projects
             }
 
             var containerClient = new BlobContainerClient(pendingUploadResponse.BlobReference.Credential.SasUri);
+
+            //var containerClient = new BlobContainerClient(
+            //    blobContainerUri: new Uri(pendingUploadResponse.Value.BlobReference.BlobUri),
+            //    credential: _tokenCredential,
+            //    options: null);
+
             return (containerClient, outputVersion);
         }
 
         /// <summary>
         /// Uploads a file to blob storage and creates a dataset that references this file.
         /// </summary>
-        public DatasetVersion UploadFile(string name, string version, string filePath, string? connectionName = null)
+        public Response<FileDatasetVersion> UploadFile(string name, string version, string filePath, string? connectionName = null)
         {
             if (!File.Exists(filePath))
             {
@@ -78,16 +84,16 @@ namespace Azure.AI.Projects
 
                 RequestContent content = new FileDatasetVersion(dataUri: dataUri).ToRequestContent();
 
-                CreateOrUpdate(name, outputVersion, content);
+                Response response = CreateOrUpdate(name, outputVersion, content);
 
-                return GetDataset(name, outputVersion);
+                return Response.FromValue(FileDatasetVersion.FromResponse(response), response);
             }
         }
 
         /// <summary>
         /// Uploads all files in a folder to blob storage and creates a dataset that references this folder.
         /// </summary>
-        public DatasetVersion UploadFolder(string name, string version, string folderPath, string? connectionName = null, Regex? filePattern = null)
+        public Response<FolderDatasetVersion> UploadFolder(string name, string version, string folderPath, string? connectionName = null, Regex? filePattern = null)
         {
             if (!Directory.Exists(folderPath))
             {
@@ -123,9 +129,9 @@ namespace Azure.AI.Projects
             Uri dataUri = uriBuilder.Uri;
 
             RequestContent content = new FolderDatasetVersion(dataUri: dataUri).ToRequestContent();
-            CreateOrUpdate(name, outputVersion, content);
+            Response response = CreateOrUpdate(name, outputVersion, content);
 
-            return GetDataset(name, outputVersion);
+            return Response.FromValue(FolderDatasetVersion.FromResponse(response), response);
         }
 
         /// <summary>
@@ -133,7 +139,7 @@ namespace Azure.AI.Projects
         /// </summary>
         private async Task<(BlobContainerClient ContainerClient, string OutputVersion)> CreateDatasetAndGetContainerClientAsync(string name, string inputVersion, string? connectionName = null)
         {
-            var pendingUploadRequest = new PendingUploadRequest(
+            PendingUploadRequest pendingUploadRequest = new(
                 pendingUploadId: null,
                 connectionName: connectionName,
                 pendingUploadType: PendingUploadType.BlobReference,
@@ -167,7 +173,7 @@ namespace Azure.AI.Projects
         /// <summary>
         /// Uploads a file to blob storage and creates a dataset that references this file.
         /// </summary>
-        public async Task<DatasetVersion> UploadFileAsync(string name, string version, string filePath, string? connectionName = null)
+        public async Task<FileDatasetVersion> UploadFileAsync(string name, string version, string filePath, string? connectionName = null)
         {
             if (!File.Exists(filePath))
             {
@@ -187,16 +193,16 @@ namespace Azure.AI.Projects
 
                 RequestContent content = new FileDatasetVersion(dataUri: dataUri).ToRequestContent();
 
-                await CreateOrUpdateAsync(name, outputVersion, content).ConfigureAwait(false);
+                Response response = await CreateOrUpdateAsync(name, outputVersion, content).ConfigureAwait(false);
 
-                return await GetDatasetAsync(name, outputVersion).ConfigureAwait(false);
+                return Response.FromValue(FileDatasetVersion.FromResponse(response), response);
             }
         }
 
         /// <summary>
         /// Uploads all files in a folder to blob storage and creates a dataset that references this folder.
         /// </summary>
-        public async Task<DatasetVersion> UploadFolderAsync(string name, string version, string folderPath, string? connectionName = null, Regex? filePattern = null)
+        public async Task<FolderDatasetVersion> UploadFolderAsync(string name, string version, string folderPath, string? connectionName = null, Regex? filePattern = null)
         {
             if (!Directory.Exists(folderPath))
             {
@@ -232,9 +238,9 @@ namespace Azure.AI.Projects
             Uri dataUri = uriBuilder.Uri;
 
             RequestContent content = new FolderDatasetVersion(dataUri: dataUri).ToRequestContent();
-            await CreateOrUpdateAsync(name, outputVersion, content).ConfigureAwait(false);
+            Response response = await CreateOrUpdateAsync(name, outputVersion, content).ConfigureAwait(false);
 
-            return await GetDatasetAsync(name, outputVersion).ConfigureAwait(false);
+            return Response.FromValue(FolderDatasetVersion.FromResponse(response), response);
         }
     }
 }